// src/modules/party-admissions/enum/admission-status.enum.ts
export enum AdmissionStatus {
    DRAFT = "DRAFT", // quần chúng đang tạo
    SUBMITTED = "SUBMITTED", // đã nộp
  
    CHI_UY_REVIEWING = "CHI_UY_REVIEWING",
    NEED_SUPPLEMENT = "NEED_SUPPLEMENT",
  
    PBT_REVIEWING = "PBT_REVIEWING",
    PBT_REJECTED = "PBT_REJECTED",
  
    LOCAL_VERIFICATION_PENDING = "LOCAL_VERIFICATION_PENDING",
    LOCAL_VERIFICATION_UPLOADED = "LOCAL_VERIFICATION_UPLOADED",
  
    RED_SEAL_CHECKING = "RED_SEAL_CHECKING",
    RED_SEAL_REJECTED = "RED_SEAL_REJECTED",
  
    RESOLUTION_DRAFTING = "RESOLUTION_DRAFTING",
    RESOLUTION_REVIEWING = "RESOLUTION_REVIEWING",
    RESOLUTION_REVISION_REQUIRED = "RESOLUTION_REVISION_REQUIRED",
  
    APPROVED = "APPROVED", // duyệt xong toàn bộ
    REJECTED = "REJECTED",
    CANCELLED = "CANCELLED",
  }