// src/modules/party-admissions/enum/admission-document-type.enum.ts
export enum AdmissionDocumentType {
    LY_LICH = "LY_LICH",
    DON_XIN_VAO_DANG = "DON_XIN_VAO_DANG",
    LY_LICH_NGUOI_XIN_VAO_DANG = "LY_LICH_NGUOI_XIN_VAO_DANG",
    GIAY_GIOI_THIEU_DANG_VIEN_1 = "GIAY_GIOI_THIEU_DANG_VIEN_1",
    GIAY_GIOI_THIEU_DANG_VIEN_2 = "GIAY_GIOI_THIEU_DANG_VIEN_2",
    NGHI_QUYET_CHI_DOAN = "NGHI_QUYET_CHI_DOAN",
    XAC_MINH_DIA_PHUONG = "XAC_MINH_DIA_PHUONG",
    NGHI_QUYET_KET_NAP_DU_THAO = "NGHI_QUYET_KET_NAP_DU_THAO",
    OTHER = "OTHER",
  }